Short circuit renderer updates on invisible elements
Speed up rendering the scene when handling invisible (opacity = 0 or visible = false) elements. Don't need to run _update methods on invisible objects. When they become visible, then run the _update method like usual. This greatly improves performance on scenes where there are a large number of elements on or off the screen and some are not visible.

This is a step towards optimizations for #363